Revert "Fix Timing Oracle in RSA decryption"
This reverts commit 8e257b86e5812c6e1cfa9e8e5f5660ac7bed899d.
